requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
openai==1.12.0
nltk==3.8.1
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1
sqlalchemy==2.0.23
aiosqlite==0.19.0
pydantic==2.5.0
python-dotenv==1.0.0
pytest==7.4.3
httpx==0.25.2
